+++ packages/components/inputs/async-select-input/src/async-select-input-open.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AsyncSelectInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const loadOptions = () =>
   Promise.resolve([
