+++ packages/components/inputs/async-select-input/src/async-select-input.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent, waitFor } from '../../../../test/test-utils';
+import { render, fireEvent, waitFor } from '../../../../../test/test-utils';
 import AsyncSelectInput from './async-select-input';
 
 // We use this component to simulate the whole flow of
