+++ packages/components/inputs/async-select-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './async-select-input';
+
+export { default as version } from './version';
