--- src/components/inputs/async-select-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './async-select-input';
